Fix database connections timing out and requiring server restarts
Previously a MySQL connection was grabbed on startup and held persistant
across all queries. While this works well for testing purposes, in a
production environment the connection is likely to time out, and since
it would go to loopback anyways connections are cheap to establish.
This commit alters the server to grab a new connection for every
instance where a database call is needed (grouped together for
performance reasons, obviously)